requirements.txt
jupyter
matplotlib
pandas
tensorflow-gpu>=1.3
tqdm
